cleanup allocator issue: no more global allocator instance move fortran bindings to `bindings/fhex`
