+++ bindings/fhex/communicator_bind.cpp
