+++ bindings/fhex/context_bind.cpp
