+++ bindings/fhex/ghex_comm_mod.f90
