+++ bindings/fhex/message_bind.cpp
@@ -1,19 +1,25 @@
-#include "message_bind.hpp"
+#include "obj_wrapper.hpp"
 #include <iostream>
 #include <cstring> 
 #include <vector>
 #include <ghex/transport_layer/message_buffer.hpp>
 #include <ghex/transport_layer/callback_utils.hpp>
+#include <ghex/allocator/unified_memory_allocator.hpp>
 
 namespace ghex = gridtools::ghex;
-
 using namespace gridtools::ghex::fhex;
 
 namespace gridtools {
     namespace ghex {
         namespace fhex {
-            // TODO how to construct the (per-thread) allocators
-            std_allocator_type std_allocator;
+
+            typedef enum 
+            {
+                 GhexAllocatorHost=1,
+                 GhexAllocatorDevice=2
+            } ghex_allocator_type;
+
+            using host_allocator_type = ghex::allocator::cuda::unified_memory_allocator<unsigned char>;
         }
     }
 }
@@ -24,16 +30,22 @@ void *ghex_message_new(std::size_t size, int allocator_type)
     void *wmessage = nullptr;
 
     switch(allocator_type){
-    case ALLOCATOR_STD:
+    case GhexAllocatorHost:
 	{
-	    ghex::tl::message_buffer<std_allocator_type> msg{size, std_allocator};
+	    ghex::tl::message_buffer<host_allocator_type> msg{size, host_allocator_type{}};
             wmessage = new ghex::tl::cb::any_message{std::move(msg)};
 	    break;
 	}
-        // case ALLOCATOR_PERSISTENT_STD:
+    case GhexAllocatorDevice:
+	{
+	    std::cerr << "BINDINGS: " << __FUNCTION__ << ": DEVICE allocator not yet implemented\n";
+            std::terminate();
+	    break;
+	}
     default:
 	{
 	    std::cerr << "BINDINGS: " << __FUNCTION__ << ": unsupported allocator type: " << allocator_type << "\n";
+            std::terminate();
 	    break;
 	}
     }
