+++ bindings/fhex/structured_bind.cpp
