+++ bindings/fhex/structured_bulk_bind.cpp
