+++ bindings/fhex/structured_staged_bind.cpp
