+++ tests/bindings/test_f_message.f90
@@ -9,7 +9,7 @@ PROGRAM test_message
   type(ghex_message) :: msg
   integer(1), dimension(:), pointer :: msg_data
   
-  msg = ghex_message_new(msg_size, ALLOCATOR_STD)
+  msg = ghex_message_new(msg_size, GhexAllocatorHost)
   
   msg_data => ghex_message_data(msg)
   msg_data(1:msg_size) = (/(i, i=1,msg_size,1)/)
